+++ src/main/java/org/example/JobSearch/model/AccountType.java
@@ -4,18 +4,18 @@ import lombok.Getter;
 
 @Getter
 public enum AccountType {
-    APPLICANT("APPLICANT", 1L),
-    EMPLOYER("EMPLOYER", 2L);
+    APPLICANT("APPLICANT", 1),
+    EMPLOYER("EMPLOYER", 2);
 
     private final String name;
-    private final Long id;
+    private final Integer id;
 
-    AccountType(String name, Long id) {
+    AccountType(String name, Integer id) {
         this.name = name;
         this.id = id;
     }
 
-    public static AccountType getById(Long id) {
+    public static AccountType getById(Integer id) {
         for (AccountType type : values()) {
             if (type.id.equals(id)) {
                 return type;
